+++ src/main/java/pvpmode/compatibility/modules/lotr/LOTRModCompatibilityModule.java
@@ -1,9 +1,10 @@
 package pvpmode.compatibility.modules.lotr;
 
-import java.io.*;
+import java.io.IOException;
 import java.lang.reflect.Field;
 import java.nio.file.*;
 import java.util.*;
+import java.util.function.Function;
 
 import cpw.mods.fml.common.FMLLog;
 import cpw.mods.fml.common.eventhandler.SubscribeEvent;
@@ -17,6 +18,7 @@ import net.minecraftforge.event.entity.player.PlayerDropsEvent;
 import pvpmode.*;
 import pvpmode.compatibility.CompatibilityModule;
 import pvpmode.compatibility.events.*;
+import pvpmode.overrides.PvPOverrideCondition;
 
 /**
  * The compatibility module for the LOTR Mod.
@@ -32,10 +34,12 @@ public class LOTRModCompatibilityModule implements CompatibilityModule
     private static final String LOTR_BIOME_IDS_FILE_NAME = "lotr_mod_biome_ids.txt";
     private static final String EXTENDED_ENEMY_BIOME_CONFIG_FILE_NAME = "extended_enemy_biomes.txt";
     private static final String DEFAULT_ENEMY_BIOME_MAP_FILE_NAME = "default_enemy_biomes_map.png";
+    private static final String SAFE_BIOME_CONFIG_FILE_NAME = "pvpmode_lotr_safe_biomes.txt";
 
     private boolean areEnemyBiomeOverridesEnabled;
     private boolean blockFTInPvP;
     private boolean dropSkullWithKeepInventory;
+    private boolean areSafeBiomeOverridesEnabled;
 
     private LOTREventHandler eventHandler;
 
@@ -52,6 +56,9 @@ public class LOTRModCompatibilityModule implements CompatibilityModule
         dropSkullWithKeepInventory = PvPMode.config.getBoolean ("Always Drop Player Skulls",
             LOTR_CONFIGURATION_CATEGORY, true,
             "If true, players killed with a weapon with the headhunting modifier will drop their skulls even with keepInventory enabled.");
+        areSafeBiomeOverridesEnabled = PvPMode.config.getBoolean ("Enable safe biome override condition",
+            LOTR_CONFIGURATION_CATEGORY, false,
+            "If true, the PvP mode safe override condition for LOTR biomes will be enabled, which has a higher priority than the enemy override condition. Players who are aligned with a faction are forced to have PvP disabled while they're in a biome which is clearly assignable to that faction. This can also be applied without the alignment criterion. This is highly configurable.");
 
         PvPMode.config.addCustomCategoryComment (LOTR_CONFIGURATION_CATEGORY,
             "Configuration entries for compatibility with the \"The Lord of the Rings Minecraft Mod\"");
@@ -59,10 +66,19 @@ public class LOTRModCompatibilityModule implements CompatibilityModule
         Path configurationFolder = PvPMode.config.getConfigFile ().getParentFile ().toPath ();
 
         FMLLog.info (String.format ("PvP mode overrides for LOTR biomes are %s",
-            areEnemyBiomeOverridesEnabled ? "enabled" : "disabled"));
+            (areEnemyBiomeOverridesEnabled || areSafeBiomeOverridesEnabled) ? "enabled" : "disabled"));
+
         if (areEnemyBiomeOverridesEnabled)
         {
             initEnemyBiomeOverrides (configurationFolder);
+        }
+        if (areSafeBiomeOverridesEnabled)
+        {
+            initSafeBiomeOverrides (configurationFolder);
+        }
+        if (areEnemyBiomeOverridesEnabled || areSafeBiomeOverridesEnabled)
+        {
+            initGeneralBiomeOverrides (configurationFolder);
         }
 
         retrieveLOTREventHandler ();
@@ -70,164 +86,50 @@ public class LOTRModCompatibilityModule implements CompatibilityModule
 
     private void initEnemyBiomeOverrides (Path configurationFolder) throws IOException
     {
-        Path enemyBiomeConfigurationFile = configurationFolder.resolve (ENEMY_BIOME_CONFIG_FILE_NAME);
+        initBiomeOverrides (configurationFolder, ENEMY_BIOME_CONFIG_FILE_NAME, "lotr enemy biome",
+            "default_enemy_biomes.txt", (data) -> new HostileBiomeOverrideCondition (data));
 
-        // Recreate the config file if it doesn't exist
-        if (!Files.exists (enemyBiomeConfigurationFile))
-        {
-            FMLLog.getLogger ().info ("The lotr enemy biome configuration file doesn't exist - it'll be created");
-            Files.createFile (enemyBiomeConfigurationFile);
-
-            PvPUtils.writeFromStreamToFile (
-                () -> this.getClass ().getResourceAsStream ("/assets/pvpmode/modules/lotr/default_enemy_biomes.txt"),
-                enemyBiomeConfigurationFile);
-        }
-
-        Map<Integer, Collection<EnemyBiomeFactionEntry>> configurationData = new HashMap<> ();
-
-        // Parse and load the config file
-        int validEntryCounter = 0;
-        int invalidEntryCounter = 0;
-        Set<String> readValidFactionEntries = new HashSet<> ();
-        try (BufferedReader reader = Files.newBufferedReader (enemyBiomeConfigurationFile))
-        {
-            // Read the file line by line
-            String line = null;
-            for (int i = 1; (line = reader.readLine ()) != null; i++)
-            {
-                line = line.trim ();
-                // Ignore comments and empty lines
-                if (!line.isEmpty () && !line.startsWith ("#"))
-                {
-                    // Split config entries into three columns
-                    String[] parts = line.split (";");
-                    if (parts.length != 3)
-                    {
-                        // There are more or less than three columns
-                        ++invalidEntryCounter;
-                        FMLLog.warning (
-                            "The lotr enemy biome config entry \"%s\" at line %d is invalid. There're too much or too less columns separated by semicolons!",
-                            line,
-                            i);
-                    }
-                    else
-                    {
-                        // Extract the faction identifier from the first column
-                        String faction = parts[0].trim ();
-                        if (faction.equals ("ALL") || LOTRFaction.forName (faction) != null)
-                        {
-                            if (readValidFactionEntries.contains (faction))
-                            {
-                                // The faction was specified already
-                                ++invalidEntryCounter;
-                                FMLLog.warning (
-                                    "The lotr enemy biome config entry at line %d references a faction (\"%s\") which was referenced by an entry loaded before. It'll be ignored.",
-                                    i, faction);
-                            }
-                            else
-                            {
-                                String alignmentString = parts[1].trim ();
-                                try
-                                {
-                                    // Extract the minimum alignment from the second column
-                                    Integer alignmentInt = Integer.parseInt (alignmentString);
-
-                                    // Extract the biome ids from the first column
-                                    String[] biomeIds = parts[2].trim ().split (",");
-                                    if (biomeIds.length <= 0)
-                                    {
-                                        // No biomes were specified
-                                        ++invalidEntryCounter;
-                                        FMLLog.warning (
-                                            "The lotr enemy biome config entry at line %d contains no assigned biome ids",
-                                            i);
-                                    }
-                                    else
-                                    {
-                                        // Parse the biome ids
-                                        Collection<Integer> biomeIdsInt = new HashSet<> ();
-                                        for (String biomeString : biomeIds)
-                                        {
-                                            String biomeStringTrimmed = biomeString.trim ();
-                                            try
-                                            {
-                                                if (!biomeIdsInt.add (Integer.parseInt (biomeStringTrimmed)))
-                                                {
-                                                    // Duplicated biome ids specified
-                                                    FMLLog.warning (
-                                                        "The lotr enemy biome config entry at line %d contains a duplicated biome id (%s).",
-                                                        i, biomeStringTrimmed);
-                                                }
-                                            }
-                                            catch (NumberFormatException e)
-                                            {
-                                                FMLLog.warning (
-                                                    "The lotr enemy biome config entry at line %d contains an invalid biome id (\"%s\"). The invalid id will be ignored.",
-                                                    i, biomeStringTrimmed);
-                                            }
-                                        }
-                                        if (biomeIdsInt.isEmpty ())
-                                        {
-                                            // Biome ids were specified, but all were invalid
-                                            ++invalidEntryCounter;
-                                            FMLLog.warning (
-                                                "The lotr enemy biome config entry at line %d contains only invalid biome ids. The entry will be ignored.",
-                                                i);
-                                        }
-                                        else
-                                        {
-                                            // Add the data to our data structured
-                                            readValidFactionEntries.add (faction);
-                                            EnemyBiomeFactionEntry entry = new EnemyBiomeFactionEntry (faction,
-                                                alignmentInt);
-                                            for (Integer biomeId : biomeIdsInt)
-                                            {
-                                                if (!configurationData.containsKey (biomeId))
-                                                {
-                                                    configurationData.put (biomeId, new HashSet<> ());
-                                                }
-                                                configurationData.get (biomeId).add (entry);
-                                            }
-                                            ++validEntryCounter;
-                                        }
-                                    }
-                                }
-                                catch (NumberFormatException e)
-                                {
-                                    ++invalidEntryCounter;
-                                    FMLLog.warning (
-                                        "The lotr enemy biome config entry at line %d contains an invalid minimum alignment (\"%s\")",
-                                        i, alignmentString);
-                                }
-                            }
-                        }
-                        else
-                        {
-                            // The faction name is invalid
-                            ++invalidEntryCounter;
-                            FMLLog.warning (
-                                "The lotr enemy biome config entry at line %d contains an invalid faction name (\"%s\").",
-                                i, faction);
-                        }
-                    }
-                }
-            }
-        }
-
-        FMLLog.info ("Loaded %d of %d specified lotr enemy biome config entries. %d config entries are invalid.",
-            validEntryCounter, invalidEntryCounter + validEntryCounter,
-            invalidEntryCounter);
-
-        PvPMode.overrideManager.registerOverrideCondition (new MiddleEarthBiomeOverrideCondition (configurationData));
-
-        // (Re)Create the LOTR biome id file
-        recreateFile (configurationFolder, LOTR_BIOME_IDS_FILE_NAME, "LOTR biome id file");
         // (Re)Create the extended enemy biome config file
         recreateFile (configurationFolder, EXTENDED_ENEMY_BIOME_CONFIG_FILE_NAME,
             "LOTR extended enemy biomes configuration file template");
         // (Re)Create the default config map image
         recreateFile (configurationFolder, DEFAULT_ENEMY_BIOME_MAP_FILE_NAME, "LOTR default enemy biome map");
+    }
 
+    private void initSafeBiomeOverrides (Path configurationFolder) throws IOException
+    {
+        initBiomeOverrides (configurationFolder, SAFE_BIOME_CONFIG_FILE_NAME, "lotr safe biome",
+            "default_safe_biomes.txt", (data) -> new SafeBiomeOverrideCondition (data));
+    }
+
+    private void initBiomeOverrides (Path configurationFolder, String configFileName, String configName,
+        String defaultConfigFileName,
+        Function<Map<Integer, Collection<BiomeFactionEntry>>, PvPOverrideCondition> conditionCreator)
+        throws IOException
+    {
+        Path biomeConfigurationFile = configurationFolder.resolve (configFileName);
+
+        // Recreate the config file if it doesn't exist
+        if (!Files.exists (biomeConfigurationFile))
+        {
+            FMLLog.getLogger ().info ("The %s configuration file doesn't exist - it'll be created", configName);
+            Files.createFile (biomeConfigurationFile);
+
+            PvPUtils.writeFromStreamToFile (
+                () -> this.getClass ().getResourceAsStream ("/assets/pvpmode/modules/lotr/" + defaultConfigFileName),
+                biomeConfigurationFile);
+        }
+
+        BiomeOverrideConfigParser parser = new BiomeOverrideConfigParser (configName,
+            biomeConfigurationFile);
+
+        PvPMode.overrideManager.registerOverrideCondition (conditionCreator.apply (parser.parse ()));
+    }
+
+    private void initGeneralBiomeOverrides (Path configurationFolder) throws IOException
+    {
+        // (Re)Create the LOTR biome id file
+        recreateFile (configurationFolder, LOTR_BIOME_IDS_FILE_NAME, "LOTR biome id file");
     }
 
     private void recreateFile (Path configurationFolder, String filename, String shortName) throws IOException